moved dataspaces for STL containers to contrib
Update #548
