+++ src/h5cpp/contrib/stl/vector.hpp
@@ -21,4 +21,30 @@ class TypeTrait<std::vector<T>> {
 };
 
 } // end of namespace datatype
+
+namespace dataspace {
+
+template<typename T>
+class TypeTrait<std::vector<T>> {
+ public:
+  using DataspaceType = Simple;
+
+  static Simple create(const std::vector<T> &value) {
+    return Simple(hdf5::Dimensions{value.size()}, hdf5::Dimensions{value.size()});
+  }
+
+  static const Dataspace & get(const std::vector<T> & value, DataspacePool & pool) {
+    return pool.getSimple(value.size());
+  }
+
+  static void *ptr(std::vector<T> &data) {
+    return reinterpret_cast<void *>(data.data());
+  }
+
+  static const void *cptr(const std::vector<T> &data) {
+    return reinterpret_cast<const void *>(data.data());
+  }
+};
+
+} // end of namespace dataspace
 } // end of namespace hdf5
